[5.8] Add package metadata to a few more ObservabilityScopes in workspace-wide and package graph loading operations (#5984) (#5986)
Clients of libSwiftPM that present diagnostics use the package metadata that can optionally be attached to an emitted Diagnostic to associate it with the package in the user interface.  This change addresses a few more cases in which no package metadata was associated.  Where possible, the metadata is associated with the ObservabilityScope at the highest place in the call stack so that any diagnostic emitted below it gets the package metadata.

rdar://103561534
(cherry picked from commit 723a5925412f756333a77fee74aedae13569814e)