build: simplify the package collection dependency graph
Convert the `PackageCollectionsModel` and `PackageCollectionsSigning`
modules static.  `PackageCollectionsModel` is used only by
`PackageCollectionsSigning` and `PackageCollections` and
`PackageCollectionsSigning` is only used by `PackageCollections`.  We
can collapse the two modules down, statically linking it into
`PackageCollections` at no cost.  This reduces the packaging complexity
as we have fewer libraries to redistribute with the toolchain.