build: convert Tool libraries into static libraries
These are used by two consumers:
- unified swift-package-manager multi-call binary
- single purpose tools

Since the single purpose tools and the multi-call binary are serving the
same purpose and will never be shipped at the same time, these libraries
can be fully internalised without increasing the size of the package. In
effect we reduce the distribution set without increasing size.