Disable implementationOnly Foundation import for resource accessor (#5997)
This was a change that landed in #5728 but it has the unintended consequence of generating unfixable warnings for packages which do import Foundation. We can probably solve that with import scanning to decide between the two import types, but given where we are in the 5.8 schedule, this seems to be something better done for a future manifest version.

(cherry picked from commit cb9d03b0aa726434ab1dd0995230e97a7bd0a09b)