Minimal handling of cancellation
If a command gets marked as cancelled from the llbuild side, we currently do not care at all and treat the build as successful which can be confusing for the user. This at least makes sure clients are notified via the delegate and the build complete message is altered.

rdar://104425290