Complete CMake build for 5.8 release (#6078)
* build: bring CMake build up to date

Add CMake based build support for the missing components as the Windows
build depends on this still.  This allows us to work towards enabling
the swift package collection and registry functionality.

* build: simplify the package collection dependency graph

Convert the `PackageCollectionsModel` and `PackageCollectionsSigning`
modules static.  `PackageCollectionsModel` is used only by
`PackageCollectionsSigning` and `PackageCollections` and
`PackageCollectionsSigning` is only used by `PackageCollections`.  We
can collapse the two modules down, statically linking it into
`PackageCollections` at no cost.  This reduces the packaging complexity
as we have fewer libraries to redistribute with the toolchain.

* build: convert Tool libraries into static libraries

These are used by two consumers:
- unified swift-package-manager multi-call binary
- single purpose tools

Since the single purpose tools and the multi-call binary are serving the
same purpose and will never be shipped at the same time, these libraries
can be fully internalised without increasing the size of the package. In
effect we reduce the distribution set without increasing size.