Add `Sendable` conformance to `Manifest` (#6099)
This is a first step towards making all model types in `PackageModel` compatible with Swift Concurrency. The only state that `class Manifest` had is `private` and is related to caching targets and dependencies. That state uses `ThreadSafeKeyValueStore`, which itself is `Sendable`.

Made `Manifest` conform to `Sendable`, since it's a model type and has no base classes or subclasses and all of its properties are declared with `let`.