Add missing `TSCUtility.Version` import (#6107)
`TSCUtility.Version` identifier is somehow implicitly re-exported in a lot of places. In some circumstances this re-exporting doesn't happen, which breaks the build.

When building `main` I've multiple times stumbled upon errors showing that `TSCUtility.Version` is not available in scope. This could only surface in certain incremental build conditions, maybe dependency scanner not allowing this identifier to be re-exported.