[NFC] PackageModel: add `Toolset` model type (#6116)
As proposed in [SE-0387](https://github.com/apple/swift-evolution/blob/main/proposals/0387-cross-compilation-destinations.md#toolsetjson-files), toolsets allow us to group paths and options for build tools. Adding this model to the codebase is a required step for fully implementing the proposal in case of its acceptance.

Added `Toolset` and `DecodedToolset` types. The latter is used for decoding from JSON, while the former is produced out of `DecodedToolset` after validation. Warnings are emitted into `ObservabilityScope` if unknown or invalid tools are found in `DecodedToolset`.

Note that it isn't actually used anywhere, so it's technically an NFC.