[NFC] Update `.swiftformat` config for consistency (#6122)
`UserToolchain.swift` was formatted by `swift-format` with a config that's inconsistent with `.swiftformat` committed in the repository, primarily in its use of 2-space indentation.

Reverted the commit that applies `swift-format`. Also updated `.swiftformat` with more tweaks that codify conventions already established in our repository. Applied new rules to `UserToolchain.swift`