[NFC] Use `@main` attribute in `swift-package-manager` target (#6127)
We'd like to use Swift Concurrency in `CrossCompilationDestinationsTool` target, but for that its commands need to conform to `AsyncParsableCommand` protocol. We have to use `@main` attribute to be able to invoke these commands from the top level.

Renamed `swift-package-manager/main.swift` to `swift-package-manager/SwiftPM.swift`, since files with `main.swift` name are always assumed to contain top-level code and `@main` attribute can't be placed in files with top-level code.