Refactor manifest serialization
This will be a fairly large change to refactor the manifest serialization to achieve two goals: get rid of the custom deserialization in `ManifestJSONParser` to use `Codable` throughout and decouple the user-facing `PackageDescription` model from its serialization.

- Use `Codable` for `ManifestJSONParser`
- Remove `Codable` conformances from public facing PD types and create a private serialization model in `PackageDescriptionSerialization` that can be used for reading and writing

Both of these have to happen in a single PR because of a few glaring issues with making the PD model fully codable.

There is further work here which can happen in subsequent PRs:

- Relocate business logic from `ManifestJSONParser` into `PackageModel` as much as possible
- Evaluate the existence of `TargetDescription` etc - it feels like we should be able to get rid of one layer of model objects here, potentially
- Evaluate other serialization code in SwiftPM, e.g. for `package describe`, the `ManifestSourceGeneration` stuff etc