Initial support for Swift macros
- Add a new module `CompilerPluginSupport` which vends the macro target type
- Represent macro targets in the package model
- Build system support for macros

In order to be able to test this end-to-end before compiler support has landed, we're building macros as dylibs using the `BUILD_MACROS_AS_DYLIBS` flag since that is already supported by the Swift compiler in `main`.

The macro API is behind the 999.0 tools version since this has not gone through Swift evolution yet.