Change executable template to use `main.swift`
The recent changes to templates switched this template away from using `@main` which seems to not be compatible with Xcode's build system. The only non-`@main` way for executables that is supported is naming the primary source file "main.swift", so that is what this PR does.

rdar://106044582

(cherry picked from commit 6fd2633da998331adeab25db4bf1df29fa11437d)