PackageModel: fix error reporting for invalid toolsets
Previously, when parsing a destination file that referenced an invalid toolset, a decoding error for the toolset was propagated as a `StringError` that mentioned the destination file, which was quite confusing. It looked as if the destination file itself couldn't be parsed, which wasn't true. Cleaning up error handling to explicitly mention the toolset file and its path if an error is thrown during its decoding.