Fix `InstallDestination.run` function not invoked by ArgumentParser (#6202)
This type uses `ParsableCommand`, not `AsyncParsableCommand`, while the `run` function itself was marked as `async`. Because of this the default implementation of `run` was always called, printing the help description even when correct arguments were passed. Removing the `async` effect from the function fixes the issue, since the function doesn't use `await` in its body anyway (due to unrelated reasons).

Also added quotes around URLs and paths in logging output so that it's easier to parse them visually, especially when those contain spaces.