Basics: explicitly mark `SQLite` as non-Sendable reference type (#6211)
`SQLite` should be a reference type as it stores a pointer to the database. For some reason it was declared as `struct`, but to indicate its reference semantics it would be cleaner to mark it as `final class`. Additionally, this type should never be `Sendable` as it provides no locking, thus an explicit annotation should be added to indicate this.