PackageModel: add DestinationConfigurationStore (#6246)
To allow users to configure installed cross-compilation destinations by providing custom paths we need to store these custom paths on the file system. 

New `DestinationConfigurationStore` allows writing out these customized properties as JSON files in a separate directory in `~/.swiftpm/destinations` or wherever that paths is symlinked to.

This is a non-functional change, new `DestinationConfigurationStore` is not used anywhere yet, but will be used in new experimental subcommands to be added in future PRs.