Basics: make `Triple` parsing errors human-readable (#6251)
Currently when a triple can't be parsed, the only error message the end user sees in the terminal is `Error: badFormat(triple: "ksjnfklsdn")`. This is quite confusing and unreadable.

Add conformance to `CustomStringConvertible` on `Triple.error`.

When a triple can't be parsed, the parsing error is surfaced to the end user in a human-readable form so that it's easier to address the error and find a root cause.