Basics: make `ObservabilityScope` conform to `Sendable` (#6089)
New `actor HTTPClient` can't take `ObservabilityScope` as an argument to its functions as it's not `Sendable`. This is the only major API difference between `HTTPClient` and `LegacyHTTPClient`, which we'd like to resolve.

Marked `SwiftToolObservabilityHandler.OutputHandler`, `ThreadSafeBox`, `ThreadSafeArrayStore`, and `ThreadSafeKeyValueStore` as `@unchecked Sendable`. Made `ObservabilityMetadata: Sendable` by requiring its keys and values to be sendable. Added additional `Sendable` requirement on `DiagnosticsHandler`.

This allows passing `ObservabilityScope` instances to `HTTPClient/execute`.