Allow sources anywhere in ./Sources when only one target is present (#6294)
After getting some feedback about the added `path` argument for executable
packages generated with `swift package init` (#6144), this change allows a
target's sources to occupy the entire "predefined sources directory" (e.g.
`Sources`, `Tests`, or `Plugins`) when there is only one target of a specific
kind occupying one of those directories in the package. All package types can
benefit from this.

For example, if there is only one library target and one test target, one can
put sources directly in `./Sources` and `./Tests`. Adding another library means
that one has to put the sources for TargetA in `./Sources/TargetA` and the
sources for TargetB in `./Sources/TargetB`, and so on.

When there is more than one target of a type in a package, the existing
requirements for target sources still apply.

This change should be compatible with existing layouts as well. If there is
only a single target in a package, then sources can of course continue to exist
exclusively in `./Sources/<target>`, or `./Tests/<testTarget>`, etc.

Amend the `executable` and `tool` template's generated manifest to not include
the `path` argument anymore as it is no longer needed.

With this change, the `library` template type can also put its sources and tests
directly in the `./Sources` and `./Tests` directories respectively.

All of the above only takes place when the tools version is set to 5.9 or
higher in the manifest.

rdar://106829666