PackageLoading: anchor relative paths for Windows
We currently cannot use `..` as a header search path for Windows even
though it may be contained within the package root.  This is due to the
fact that the path itself was unanchored, and a raw `..` would be
treated as the drive root which will almost always be outside of the
package root.  Construct the absolute path relative to the target root
rather than appending, permitting us to form a well-formed path which
can be checked for ancestry.