Expected signing entity verification
This allows clients to pass in a dictionary with expected signing entities that SwiftPM will check after loading the package graph. This can be used by clients to provide a priori configuration of expected signing by the user or provide a way to verify that information that was previously shown to users matches what was verified during signature verification.

Note that since this operates at the workspace level, we're verifying against the data cached during signature verification, not against the actual data.

rdar://107162424