Add templates for build tool plugins and command plugins (#6111)
Adds templates for build tool plugins and command plugins, allowing them to be created from the `swift package init` command.  The packages are set up to vend the plugins via plugin product declarations, so they can be used by other packages.  The initial code for each each plugin type is minimal, but for the build tool plugin type it shows how to create and return a build command for each source file of a particular type, since that's the most common use of build tool plugins.

rdar://107288352

Co-authored-by: Anders Bertelrud <anders@apple.com>