[5.9] Cross-compilation: fix bundles not unpacked on installation (#6362)
`swift experimental-destination install` subcommand works with bundle directories, but fails to unpack bundle archives.

We should pass an instance of an archiver to the installation function to unpack destination bundle archives if needed. Since `withTemporaryDirectory` is `async`, `SwiftDestinationTool` had to be converted to `AsyncParsableCommand` and also gain `@main` attribute for that to work. Additionally, `@main` attribute requires `-parse-as-library` passed in CMake.

rdar://107367895