Fix product lookup
For module aliasing, we changed the product lookup to include package identity, but product references can use a local identity as specified by the `name` attribute of the package dependency declaration. So for packages where the local `name` attribute didn't match the identity, we would not longer find the corresponding products at all. This switches us back to use the local identity that is also called `explicitNameForTargetDependencyResolutionOnly` in package reference.

rdar://106578471