[5.9] Cross-compilation: fix support for `.tar.gz` bundles (#6393)
rdar://107485048

`swift experimental-destination install` subcommand fails when pointed to a destination bundle compressed with `.tar.gz` extension, compressed with tar and gzip. This archival format is vastly superior to zip in the context of cross-compilation destination bundles. In our typical scenario we see that `.tar.gz` archives are at least 4x better than `.zip`. Since in absolute terms this amounts to multiple gigabytes, this has substantial impact on the amount of time and traffic taken by downloading and installing such bundle archives.

Added `TarArchiver` and `UniversalArchiver` types that support `.tar.gz` archives. `UniversalArchiver` is now used in `InstallDestination` instead of `ZipArchiver`.

Risk: low, since the change is isolated to the cross-compilation destinations feature and is covered with tests.