Re-enable disabled tests
This re-enables a bunch of disabled tests, the hope is that they should now work that we have a newer superior Xcode installed on CI. If that happens to not be the case, we have `SWIFTCI_DISABLE_SDK_DEPENDENT_TESTS` to disable these tests specifically for the macOS CI if needed. This change introduces the variable where expected to be needed, but doesn't set it.

rdar://96920453