Rename "CC Destination" to "Swift SDK"
"Cross-compilation destination" is a very verbose name. "Swift SDK" (note that this is distinct from just plain "SDK" used in other contexts) is a nicer alternative. Not all of the types and functions were renamed yet, especially `Destination`, which has been present in the codebase for long time. Renaming that type will be potentially source-breaking for libSwiftPM clients. This change is focused and updating user-visible strings and renaming types and functions that were only recently added. CLI subcommands were renamed accordingly.