Re-enable disabled tests
* Re-enable disabled tests

This re-enables a bunch of disabled tests, the hope is that they should now work that we have a newer superior Xcode installed on CI. If that happens to not be the case, we have `SWIFTCI_DISABLE_SDK_DEPENDENT_TESTS` to disable these tests specifically for the macOS CI if needed. This change introduces the variable where expected to be needed, but doesn't set it.

* Removed back deployment rpaths
* Removed more Darwin rpaths into the toolchain

A macOS that contains the dyld issue worked around by https://github.com/apple/swift/pull/37978 has been deployed by Swift CI, so we can't do this anymore on Darwin.

* Do not include Swift runtime libs in `DYLD_LIBRARY_PATH` on Darwin

(cherry picked from commit 16fb1d2c208bea343ad71e66fd1969bdccf11f47)