Basics: remove `AsyncFileSystem` as unused (#6436)
Since `FileSystem` is now `Sendable`, we no longer need to keep `AsyncFileSystem`, which served only as a workaround for non-sendability. It wasn't non-blocking under the hood anyway, and a blocking FS I/O API would have to be designed from scratch anyway, not as a wrapper for the existing blocking APIs.

`AsyncFileSystem` actor declaration and its uses were removed. `FileSystem` can be used directly in our async code instead.