Change `testPluginCanBeReferencedByProductName`
We have recently seen some failures in this test e.g. https://ci.swift.org/job/oss-swift-package-ubuntu-20_04/1779/ -- it is not clear to me why these failures started to appear, but the diagnostic message that seems to be missing is not actually material to the test, so we can change the test to avoid this.

rdar://109368270

(cherry picked from commit 96db9f664e7347c403192bd6718435f2a30df69f)