New build tool plugin and command plugin templates should have conditional support for XcodeProjectPlugin (#6446)
Many package plugins are used from Xcode, so it makes sense for newly created ones to have that by default.  If there is support for other IDEs in the future, it would make sense to add that as well.

This support is currently always added but we could consider having options in the API for whether to add it.

rdar://108166582
(cherry picked from commit f0eae334054823e5a75ae45ffa3f6f83ae72fbac)