Support macro targets in plugin context
This currently prevents the use of plugins in any package graph that happens to include one or more macro targets which seems extremely unfortunate.

This fix is slightly ugly because `PackagePlugin` should not concretely know about macros. An alternative could be excluding macros from the project model that plugins see.

rdar://108531004