Fix use of XCTest minimum deployment target
If the computed XCTest minimum deployment target is lower than the oldest supported version of a platform, we should not use it as the basis for deployment targets of test targets.

This issue was actually also present in the tests which were written to expect the incorrect behavior of always preferring the XCTest minimum deployment target.

rdar://108462045

(cherry picked from commit 8af71a96ca22e13b32fa64e2c2bc86826fd9fe41)