Basics: introduce some native path representation helpers (#6454)
We end up needing these in a few places, and while we could make it
`internal` and use `@testable` imports in the tests, this ends up being
needed in `Build` and `PackageLoading`.  Ideally, we would give it
package visibility, but that functionality would not be backwards
compatible.  This is a reluctant approach at solving that problem by
using `public` visibility.