Allow `--product` option for automatic library products
This is currently explicitly disallowed because we're not producing a binary for these products, but that seems like an odd restriction since `--target` also only produces intermediates. It seems like it can be useful to build a subset of the package for other reasons than wanting to produce a particular binary.