SourceControl: centralise git cloning (#6550)
* SourceControl: remove outdated FIXME (NFC)

The `progresHandler` is the callback and specifying it will attach
`--progress` to the invocation and will perform the callback.

* SourceControl: run through SwiftFormat (NFC)

* SourceControl: centralise git clone logic

Create a helper to invoke `git clone` since we pass a few options to the
invocation.  This ensures that we always clone a repository with the
same set of options and that we do not need to keep the various sites in
sync.