Revert "missing pinned version" change for 5.10
There's an unfortunate interaction here with the resolver precomputation / `LocalPackageContainer` which leads to us rejecting versions as "not available" if they don't happen to be the currently checked out version. We need a more hollistic change to resolve this issue which is out of scope for 5.9, so we'll revert this change here.

rdar://115303494

(cherry picked from commit 87d65e3da6a128b87f105f579caa58e509962617)