Revert "missing pinned version" change for main
There's an unfortunate interaction here with the resolver precomputation / `LocalPackageContainer` which leads to us rejecting versions as "not available" if they don't happen to be the currently checked out version. We need a more hollistic change to resolve this issue which is out of scope for 5.9, so we'll revert this change here.

We need to cherry-pick this to main to make the main => 5.10 automerger viable.

(cherry picked from commit 87d65e3da6a128b87f105f579caa58e509962617)