Build: migrate to file lists for driver invocations
When invoking the driver use file lists rather than passing all the
sources on the command line. This is particularly import for Windows
where the command line limit is 32k. With deep dependency sets and large
projects, it is possible to overrun this limit and fail to invoke the
command. The driver itself will invoke the frontend using response files
if the invocation will overflow the command line limit.

Fixes: #6402