Tests: attempt to handle arc separators in tests
The arc separator on Windows is `\` rather than `/`.  The path tests
hard coded the expectation that the unix path separator is always used
which is not a valid assumption.  Add a branch for windows in a large
number of cases.  Although we do not fully pass the tests, this gets us
closer.