Workspace,Build: reduce `mtime` updates
When building a project inside of VSCode with the LSP enabled,
rebuilding the project would trigger a change to the mtimes of the
auxiliary files resulting in an unnecessary rebuilding of targets.  In
conjunction with the changes to tools-support-core
(apple/swift-tools-support-core#422), incremental builds do not trigger
a cascade of rebuilds that are unnecessary.

Extracted from a patch by Ami Fischman!

Co-authored-by: Ami Fischman <ami@thebrowser.company>