Add `visionOS` as a platform alongside `iOS`, `tvOS`, `watchOS`, and the others (#6662)
Add new enum constants for visionOS in all the places where it is appropriate.

This includes the `platforms` parameter in the manifest, whose purpose is to list the minimum deployment targets that a package requires. Since there is initially only visionOS 1.0, and because absence of a platform in the `platforms` parameter means that the default version is supported (not that the platform is unsupported), all packages implicitly support visionOS 1.0 whether or not they list it in the `platforms` parameter. But having the enum constant alongside those for the other Apple platforms makes everything consistent, and will become important for any future versions.

This change also makes the `visionOS` platform available for use in conditional target dependencies.

rdar://107064954
(cherry picked from commit 809cb152a303926b243edb09ba4f0590e4b0c8b7)