Build: centralise the debug information flag handling
Rather than spreading out the flag addition across the various different
targets, we can control the flags at the `BuildParameters` level,
driving the necessary flag addition at a single site.  This would allow
easier management of the flags as we can see the additions uniformly
applied.

This shuffles the previously introduced functionality from `SwiftTool`
and sinks it into `BuildParameters`.  This enables the application of
the flags without having to go through the tooling, i.e. when SPM is
used as a library.

The test changes alter the expectations for ordering and additional
flags being applied for various compilation cases.