Remove `swiftpm-xctest-helper`
A while ago, the canonical version of the helper tool was moved to Xcode itself, but in OSS toolchains and local builds, we were still using the prior version of the tool. This removes the previous tool and adjusts the code for finding the correct tool in the user's Xcode installation.

rdar://74658052