Update Swift SDK nomenclature (#6678)
Introduced `SwiftSDKMetadataV4` type that uses `targetTriple` instead of `runTimeTriple`. We're keeping the previous v3 schema for better error handling in case v3 schema ended up being used in the wild during Swift 5.9 development cycle. The plan is to deprecate v3 and earlier schemas in a future change.

`buildTimeTriple` and `runTimeTriple` renamed to `hostTriple` and `targetTriple` across the codebase, destinations renamed to Swift SDKs where possible without a broader disruption. Full renaming of the `Destination` type to `SwiftSDK` would come in a separate change to reduce the diff for this one.