Remove extra paths added to LD_LIBRARY_PATH on Linux when running tests
These break running tests when there's already a swift on PATH that
isn't the swift currently being run. There shouldn't be a need for these
as the run path of the compiled binary is already the correct path (ie.
the library path of the swift that built it and `$ORIGIN`).

(cherry picked from commit 5c7db5876c0e88d7bb8be97620131d8be7921713)