CoreCommands: support `-debug-info-format none` for Windows
Windows does not have a reliable way to strip the binary as stripping is
not an operation that MSVC ever performs (PDBs are always externalised
and have support for private and public symbols).  This allows us to
have builds in release mode which do not embed debug information which
would then need to be extricated.