Fix Swift SDK handling URLs without filename components
Currently when invoking `swift experimental-sdk install https://server.org/abc?xy=z` installation will fail as we assume URLs always have a filename component similar to `https://server.org/swift-sdk.tar.gz`. Let's make the installation logic more flexible so that bundle name is inferred from the downloaded extract archive contents and not the URL.

Also improved logging so that final bundle name is printed for such URLs without a filename component. Additionally we now print a message when starting a downloading, which is helpful if a download is long-running.