[NFC] Rename `Destination` type to `SwiftSDK` (#6707)
With the cross-compilation proposal accepted in principle, it's time to move over to the new nomenclature altogether.

Renamed `Destination` type to `SwiftSDK`. Also renamed mentions of "destination" in declarations to "Swift SDK" and cleaned up usage of `triple` where it was actually referring to the target triple.

Naming is consistent across the codebase. This is an NFC.

* Rename `Destination` type to `SwiftSDK`

Also renamed more mentions of "destination" in declarations to "Swift SDK".

* Fix build error with incomplete renaming

* Add `@_disfavoredOverload` to deprecated `BuildParameters.init`