Fix CMake build to explicitly pick up ar and ranlib from toolchain.
CMake under various conditions will attempt to use `llvm-ar` and
`llvm-ranlib` when it spots that we're building using `clang`, and
while that works fine on Linux, it doesn't work so well on Darwin,
leaving us with a link failure with a perplexing error message.

The rest of the Swift build is resilient to this problem already,
but the `bootstrap` script here doesn't pass the relevant options
to CMake, so it's possible to end up with the build failing when
trying to link `lib/libCYaml.a` after the build has mangled the
static library with `llvm-ranlib`.

rdar://112478653