Add `SerializedJSON` type as string literal escaping helper (#6730)
Currently we have to call `_nativePathString` manually in a lot of places where `AbsolutePath` is interpolated into strings. We also need to take into account whether `_nativePathString` should be called with `escaped: true` or `escaped: false` arguments, which is prone to errors. The serialization format should carry the knowledge of its correct escaping algorithm, not the path representation itself.

We can make this much cleaner if string interpolations call `_nativePathString` automatically where needed, and serialization formats add proper escaping on top of that.

Let's mplements this for JSON, specifically where it is used in Swift SDKs.

### Modifications:

Added `SerializedJSON` type as string literal escaping helper.

Also added a custom `AbsolutePath` interpolation to the new type, which removes the need to call `_nativePathString` when paths are interpolated in JSON strings.

### Result:

We don't have to call `_nativePathString` manually in as many places. In future PRs `_nativePathString` should be declared `private` or at least `internal` so that it's hidden as an implementation detail of appropriate string interpolations.